refactor: centralize dtype handling via DtypePolicy (#18)
* build: bump scikit-image for Python 3.12 and disable ffcv (no wheel, missing system deps)

* refactor: centralize dtype handling via DtypePolicy

Fix the test-time dtype mismatch (Half input vs float32 conv) by making
TrainerParams.precision the single source of truth for dtype resolution.

- Add dynvision/utils/dtype_policy.py: canonical PRECISION_TO_DTYPE map
  (stored-parameter semantics), resolve_dtype (warn + float32 default),
  DtypePolicy dataclass, and shared coordinate_component_dtypes helper.
- Route get_effective_dtype_from_precision and DtypeDeviceCoordinator
  through the shared resolver; replace silent float16 fallbacks with
  warn + float32.
- Delete DataParams.precision and dead ModelParams.target_dtype.
- Apply the same coordinate_component_dtypes validator to TestingParams
  (removing the 'precision' alias that swallowed unscoped precision and
  caused data to derive float16).
- Default dataloaders to float32.

Regression tests in tests/params/test_dtype_policy.py.

* fix(dtype): correct bfloat16 mapping, effective_dtype lookup, and optional ffcv import

- map_dtype() in coordination.py stripped the substring "float" from all
  spellings, turning "bfloat16"/"torch.bfloat16" into "b16", which
  resolve_dtype() couldn't match and silently fell back to float32.
  Now bfloat16 spellings normalize to "bf16" explicitly before the
  "float" strip is applied to the other spellings.

- DataParams.effective_dtype did a string-keyed dict lookup on self.dtype,
  but validate_dtype already converts self.dtype to torch.dtype, so the
  lookup always missed and silently returned torch.float32 regardless of
  the configured dtype. Now returns self.dtype directly.

- data_params.py and transforms.py had unconditional top-level
  `from ffcv.loader import OrderOption` / `import ffcv.transforms`, but
  ffcv is commented out of pyproject.toml (no wheel / missing system
  deps). This broke every test module that imports dynvision.params or
  dynvision.data.transforms in environments without ffcv installed
  (8 CI import errors). Both imports are now optional: data_params.py
  falls back to a local OrderOption-mirroring IntEnum, and transforms.py
  raises a clear ImportError only if backend="ffcv" is actually
  requested without ffcv installed.

Adds regression tests for both dtype bugs and verifies the ffcv-optional
import path by blocking ffcv at runtime.

* test(transforms): skip ffcv-backend tests when ffcv is not installed

ffcv is an optional dependency (commented out of pyproject.toml — no
wheel / missing system deps), but tests/data/test_transforms.py had no
guard for the tests exercising backend="ffcv". After making the ffcv
imports in dynvision/data/transforms.py and dynvision/params/data_params.py
optional, these tests started failing loudly in CI with the new, correct
ImportError instead of the previous collection-time ModuleNotFoundError
that masked the whole test module.

Add a requires_ffcv skip marker (based on actual ffcv importability) to
the three tests that genuinely need a working ffcv install:
test_parse_bare_module_name_ffcv, test_get_ffcv_train_transforms,
test_ffcv_mnist_train_workflow. Verified clean pass+skip both with ffcv
present and with it blocked at import time.

* Refactor dtype alignment logic in dtype_policy.py


---------

+++ dynvision/data/ffcv_dataloader.py
@@ -34,7 +34,7 @@ def _build_image_pipeline(
     data_transform: Optional[List[Callable]],
     n_timesteps: int = 0,
     normalize: Optional[Tuple[List, List]] = None,
-    dtype: Optional[torch.dtype] = torch.float16,
+    dtype: Optional[torch.dtype] = torch.float32,
     device: Optional[torch.device] = None,
 ) -> List[Operation]:
     if encoding == "tensor":
@@ -113,7 +113,7 @@ def get_ffcv_dataloader(
     encoding: str = "image",
     resolution: int = 224,
     drop_last: bool = True,
-    dtype: Optional[torch.dtype] = torch.float16,
+    dtype: Optional[torch.dtype] = torch.float32,
     batches_ahead: int = 2,
     train: bool = True,
     verbose: bool = False,
